# docs/swagger.yaml
definitions:
  model_base.BaseArticle:
    properties:
      IsOpen:
        description: 是否公开文章
        type: string
      articleId:
        description: 文章id
        type: integer
      articleUrl:
        description: 文章路径
        type: string
      content:
        description: 内容
        type: string
      groupId:
        description: 组id
        type: integer
      labelIds:
        description: 标签
        type: string
      libraryId:
        description: 库id
        type: integer
      parentId:
        description: 父级id
        type: integer
      status:
        description: 状态
        type: string
      title:
        description: 标题
        type: string
      typeId:
        description: 类型
        type: integer
      userId:
        description: 用户id
        type: integer
    type: object
...
  result.Result:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  token.LoginForm:
    properties:
      ak:
        description: 账号（必填）
        type: string
      ck:
        description: 验证码（登录时填写图片验证码，短信和邮箱则填写发送的验证码）
        type: string
      ckId:
        description: 验证码id（验证码对应的id索引）
        type: string
      lt:
        description: 登录类型（必填）1账号密码，2短信验证码，3邮箱验证码，4微信扫码
        type: string
      pk:
        description: 密码（账号密码登录必填）
        type: string
      rkId:
        description: rsa秘钥id（加密传输的秘钥id索引）
        type: string
    type: object
  token.RegisterForm:
    properties:
      ck:
        type: string
      ckId:
        type: string
      cpk:
        type: string
      mobile:
        type: string
      pk:
        type: string
      rkId:
        type: string
      uk:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: loginForm
        in: body
        name: loginForm
        required: true
        schema:
          $ref: '#/definitions/token.LoginForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 登录接口（无需加密）
      tags:
      - Auth - 认证授权
  /auth/loginKey:
    post:
      consumes:
      - application/json
      parameters:
      - description: loginForm
        in: body
        name: loginForm
        required: true
        schema:
          $ref: '#/definitions/token.LoginForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 登录接口（加密）
      tags:
      - Auth - 认证授权
  /auth/logout:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 退出登录
      tags:
      - Auth - 认证授权
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: registerForm
        in: body
        name: registerForm
        required: true
        schema:
          $ref: '#/definitions/token.RegisterForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 登录接口（无需加密）
      tags:
      - Auth - 认证授权
  /auth/registerKey:
    post:
      consumes:
      - application/json
      parameters:
      - description: registerForm
        in: body
        name: registerForm
        required: true
        schema:
          $ref: '#/definitions/token.RegisterForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 登录接口（加密）
      tags:
      - Auth - 认证授权
  /base/baseArticle/{articleId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: articleId
        in: path
        name: articleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取文章详情
      tags:
      - Base - 文章管理
  /base/baseArticle/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: baseArticle
        in: body
        name: baseArticle
        required: true
        schema:
          $ref: '#/definitions/model_base.BaseArticle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 添加文章
      tags:
      - Base - 文章管理
  /base/baseArticle/deletes:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: articleIds
        in: query
        name: articleIds
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 删除文章（实体删除）
      tags:
      - Base - 文章管理
  /base/baseArticle/edit:
    put:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: baseArticle
        in: body
        name: baseArticle
        required: true
        schema:
          $ref: '#/definitions/model_base.BaseArticle'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 修改文章
      tags:
      - Base - 文章管理
  /base/baseArticle/myPage:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 是否公开文章
        in: query
        name: IsOpen
        type: string
      - description: 文章id
        in: query
        name: articleId
        type: integer
      - description: 文章路径
        in: query
        name: articleUrl
        type: string
      - description: 内容
        in: query
        name: content
        type: string
      - description: 开始时间
        in: query
        name: endTime
        type: string
      - description: 组id
        in: query
        name: groupId
        type: integer
      - description: 标签
        in: query
        name: labelIds
        type: string
      - description: 库id
        in: query
        name: libraryId
        type: integer
      - description: 当前页（非分页接口不传）
        in: query
        name: pageNum
        type: integer
      - description: 每页条数（非分页接口不传）
        in: query
        name: pageSize
        type: integer
      - description: 父级id
        in: query
        name: parentId
        type: integer
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 标题
        in: query
        name: title
        type: string
      - description: 类型
        in: query
        name: typeId
        type: integer
      - description: 用户id
        in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取当前用户的文章列表
      tags:
      - Base - 文章管理
  /base/baseArticle/page:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: 是否公开文章
        in: query
        name: IsOpen
        type: string
      - description: 文章id
        in: query
        name: articleId
        type: integer
      - description: 文章路径
        in: query
        name: articleUrl
        type: string
      - description: 内容
        in: query
        name: content
        type: string
      - description: 开始时间
        in: query
        name: endTime
        type: string
      - description: 组id
        in: query
        name: groupId
        type: integer
      - description: 标签
        in: query
        name: labelIds
        type: string
      - description: 库id
        in: query
        name: libraryId
        type: integer
      - description: 当前页（非分页接口不传）
        in: query
        name: pageNum
        type: integer
      - description: 每页条数（非分页接口不传）
        in: query
        name: pageSize
        type: integer
      - description: 父级id
        in: query
        name: parentId
        type: integer
      - description: 开始时间
        in: query
        name: startTime
        type: string
      - description: 状态
        in: query
        name: status
        type: string
      - description: 标题
        in: query
        name: title
        type: string
      - description: 类型
        in: query
        name: typeId
        type: integer
      - description: 用户id
        in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 获取文章列表
      tags:
      - Base - 文章管理
  /base/baseArticle/removes:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - collectionFormat: csv
        description: articleIds
        in: query
        items:
          type: integer
        name: articleIds
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/result.Result'
      summary: 删除文章（逻辑删除）
      tags:
      - Base - 文章管理
...
swagger: "2.0"